+++ web-app/package.json
@@ -22,7 +22,6 @@
     "axios": "^1.7.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-google-autocomplete": "^2.7.3",
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.2",
     "@windy-civi/domain": "file:../web-app",
