+++ web-app/public/privacy.html
@@ -58,10 +58,6 @@
     <p>We collect the following types of information when you use our App:</p>
     <ul>
       <li>
-        <strong>Address</strong>: We use your address to find your
-        representatives. We do not store your address anywhere in our servers.
-      </li>
-      <li>
         <strong>Data Collected Automatically</strong>: We don't do any high
         fidelity analytics tracking of our users. There's no tracking cookies or
         similar.
