+++ web-app/src/app/config.ts
@@ -1,7 +1,9 @@
 import { Env } from "@windy-civi/domain/types";
 
-export const getEnv = (env: ImportMetaEnv): Env => {
-  return {
-    GOOGLE_API_KEY: env.VITE_GOOGLE_API_KEY,
-  };
+export const getEnv = (
+  // Don't need this, but keeping in case we need it in the future
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  env: ImportMetaEnv,
+): Env => {
+  return null;
 };
