+++ web-app/src/app/feed/loader.tsx
@@ -1,10 +1,9 @@
 import { json, type LoaderFunction } from "react-router-dom";
 import { getEnv } from "../config";
 
-import { getFilteredLegislation } from "@windy-civi/domain/filters/filters.api";
+import { getFeed } from "@windy-civi/domain/feed/api";
 
 import { viteDataGetter } from "../../api/vite-api";
-import { getRepresentativesWithCache } from "@windy-civi/domain/representatives/representatives.api";
 import { getPreferencesFromCookies } from "../preferences/api";
 import { type FeedLoaderData } from "./types";
 
@@ -12,19 +11,14 @@ export const loader: LoaderFunction = async () => {
   const env = getEnv(import.meta.env);
   const userPreferences = await getPreferencesFromCookies(document.cookie);
 
-  const representatives = await getRepresentativesWithCache(
-    env,
-    userPreferences.location,
-  );
-  const feedData = await getFilteredLegislation({
-    representatives,
-    filters: userPreferences,
+  const feedData = await getFeed({
+    preferences: userPreferences,
     dataStoreGetter: viteDataGetter,
   });
 
   return json<FeedLoaderData>({
     env,
     userPreferences,
-    feedData: feedData.filteredLegislation,
+    feed: feedData.feed,
   });
 };
