+++ web-app/src/app/feed/types.ts
@@ -3,5 +3,5 @@ import { Env, UserPreferences, WindyCiviBill } from "@windy-civi/domain/types";
 export interface FeedLoaderData {
   env: Env;
   userPreferences: UserPreferences;
-  feedData: WindyCiviBill[];
+  feed: WindyCiviBill[];
 }
