+++ web-app/src/app/preferences/loader.tsx
@@ -1,9 +1,7 @@
 import { json, type LoaderFunction } from "react-router-dom";
 import { getEnv } from "../config";
 
-import { parseAvailableTags } from "@windy-civi/domain/filters/filters.utils";
-import { getRepresentativesWithCache } from "@windy-civi/domain/representatives/representatives.api";
-import { getAllOffices } from "@windy-civi/domain/representatives/representatives.utils";
+import { parseAvailableTags } from "@windy-civi/domain/feed/utils";
 import { UserPreferencesLoaderData } from "./types";
 import { getPreferencesFromCookies } from "./api";
 
@@ -15,7 +13,6 @@ import { getPreferencesFromCookies } from "./api";
 
 // Other Saved Data
 // -- Available Tags
-// -- Representatives
 
 export const loader: LoaderFunction = async () => {
   // Env comes from vite
@@ -24,19 +21,11 @@ export const loader: LoaderFunction = async () => {
   // Preferences are in cookies
   const preferences = await getPreferencesFromCookies(document.cookie);
 
-  // Saved data, like feed sources, representatives, and available tags are in a local localStorage.
-  // Todo: move this to getting from IndexedDB
-  const representatives = await getRepresentativesWithCache(
-    env,
-    preferences.location,
-  );
-  const offices = getAllOffices(representatives);
   const availableTags = parseAvailableTags(preferences.location);
 
   return json<UserPreferencesLoaderData>({
     env,
     data: {
-      offices,
       availableTags,
     },
     preferences,
