+++ web-app/src/app/preferences/types.ts
@@ -1,10 +1,8 @@
-import { OfficialOffice } from "@windy-civi/domain/representatives/representatives.types";
 import { Env, UserPreferences } from "@windy-civi/domain/types";
 
 export type UserPreferencesLoaderData = {
   env: Env;
   data: {
-    offices: OfficialOffice[] | null;
     availableTags: string[];
   };
   preferences: UserPreferences;
